Makefile: Remove '-nostdlib' option from the linker flags
The `-nostdlib` option does not add any optimization in code size compared to
code generated without this option, this has been tested as seen in the following
issue https://github.com/Optiboot/optiboot/issues/233
But this does generate linker errors of undefined functions such as
`eeprom_write_byte()` and `eeprom_read_byte()` when the option is added to
the `LDFLAGS`. This problem did not exist in older avr-libc versions, where these
functions were in libc(which was explicitly included in the link with '-lc'), but
in newer versions these functions were moved into an mcu-specific library which
is excluded from the build with `-nostdlib`.
Removing the option '-nostdlib' fixes this issue without changing the compiled
hex files in any way.