Bump "OPTIBOOT_CUSTOMVER" for custom Devoboard version builds
Since adding Devoboard versioning to the Optiboot bootloader makes
the bootloaders a custom build, we must bump the "OPTIBOOT_CUSTOMVER"
to comply with the original Optiboot versioning.
Bumping the "OPTIBOOT_CUSTOMVER" also enables us to see if the Devoboard
versions are added to the bootloader, else we would get random memory readouts
at the same ".version" section address.