Mark versions struct as const
This allows the compiler to inline the contained values in the
`STK_GET_PARAMETER` handling. Without this, these version variables are
read from RAM at runtime, but their values are only present in flash,
not RAM, so garbage data is returned. In particular, before this commit,
avrdude would show:

         Hardware Version: 128
	 Firmware Version: 129.130

(though these values might be dependent on how avrdude is called,
depending on what values from RAM are read exactly)

Note that this relies on the compiler to do proper optimization. It
might be more reliable to instead read these variables with
pgm_read_byte instead, but to minimize the difference with upstream
optiboot, this is not done now.